+++ R/fn.R
@@ -474,7 +474,9 @@ as_closure <- function(x, env = caller_env()) {
       names(args)[(names(args) == "...")] <- ""
 
       prim_call <- call2(fn_name, splice(args))
-      new_function(fmls, prim_call, base_env())
+      # The closure wrapper should inherit from the global environment
+      # to ensure proper lexical dispatch with methods defined there
+      new_function(fmls, prim_call, global_env())
     }
   )
 }